Feat: fix the module data.py
issue #63
