--- backend/app/db/Data.py
@@ -1,4 +1,0 @@
-from sqlalchemy import create_engine
-
-url_data = "sqlite:///./data"
-engine = create_engine(url_data)
